Even more changes, hopefully that's it
