+++ samples/snippets/csharp/VS_Snippets_CLR/conceptual.disposable/cs/Bar.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+public sealed class Bar : IDisposable
+{
+    public void Dispose()
+    {
+    }
+}
