+++ samples/snippets/csharp/VS_Snippets_CLR/conceptual.disposable/cs/Disposable.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+public class Disposable : IDisposable
+{
+    // <SnippetDispose>
+    public void Dispose()
+    {
+        // Dispose of unmanaged resources.
+        Dispose(true);
+        // Suppress finalization.
+        GC.SuppressFinalize(this);
+    }
+    // </SnippetDispose>
+
+    // <SnippetDisposeBool>
+    protected virtual void Dispose(bool disposing)
+    {
+    }
+    // </SnippetDisposeBool>
+}
