+++ samples/snippets/csharp/VS_Snippets_CLR/conceptual.disposable/cs/using.csproj
@@ -4,6 +4,7 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
     <Nullable>enable</Nullable>
+    <StartupObject>SingleStatementMultiple</StartupObject>
   </PropertyGroup>
 
 </Project>
