+++ samples/snippets/visualbasic/VS_Snippets_CLR/conceptual.disposable/vb/Bar.vb
@@ -1,0 +1,6 @@
+﻿Public NotInheritable Class Bar
+    Implements IDisposable
+
+    Public Sub Dispose() Implements IDisposable.Dispose
+    End Sub
+End Class
