+++ samples/snippets/visualbasic/VS_Snippets_CLR/conceptual.disposable/vb/using.vb.vbproj
@@ -4,6 +4,7 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
     <Nullable>enable</Nullable>
+    <StartupObject>using.vb.TryExplicitCatchFinally</StartupObject>
   </PropertyGroup>
 
 </Project>
